refactor(hfb): moved HFBReader to containers
