+++ ch_pipeline/hfb/__init__.py
@@ -7,6 +7,5 @@ CHIME HFB Pipeline
     :toctree:
 
     containers
-    io
 
 """
